Move eval server to a module file
